+++ src/main/java/codesquad/http/parser/HttpRequestParser.java
@@ -1,0 +1,37 @@
+package codesquad.http.parser;
+
+import codesquad.http.HttpRequest;
+import codesquad.http.type.StartLine;
+
+import java.util.Map;
+
+public class HttpRequestParser {
+    private HttpRequestParser() {}
+
+    public static HttpRequest parse(String message) {
+        String[] lines = message.split("\n");
+
+        StartLine startLine = StartLineParser.parse(lines);
+        Map<String, String> headers = HeaderParser.parse(lines);
+
+        // Body
+        int contentLength = getContentLength(headers);
+        String body = BodyParser.parse(lines, contentLength);
+
+        return new HttpRequest(
+                startLine.method(),
+                startLine.path(),
+                startLine.query(),
+                startLine.protocol(),
+                headers,
+                body);
+    }
+
+    private static int getContentLength(Map<String, String> headers) {
+        int contentLength = 0;
+        if (headers.containsKey("Content-Length")) {
+            contentLength = Integer.parseInt(headers.get("Content-Length"));
+        }
+        return contentLength;
+    }
+}
