+++ src/main/java/codesquad/http/type/StartLine.java
@@ -1,0 +1,11 @@
+package codesquad.http.type;
+
+import java.util.Map;
+
+public record StartLine(
+        HttpMethod method,
+        String path,
+        Map<String, String> query,
+        HttpProtocol protocol
+) {
+}
